+++ actix-http/src/h1/client.rs
@@ -1,4 +1,4 @@
-use std::io;
+use std::{fmt, io};
 
 use actix_codec::{Decoder, Encoder};
 use bitflags::bitflags;
@@ -17,9 +17,9 @@ use crate::{
 
 bitflags! {
     struct Flags: u8 {
-        const HEAD              = 0b0000_0001;
-        const KEEPALIVE_ENABLED = 0b0000_1000;
-        const STREAM            = 0b0001_0000;
+        const HEAD               = 0b0000_0001;
+        const KEEP_ALIVE_ENABLED = 0b0000_1000;
+        const STREAM             = 0b0001_0000;
     }
 }
 
@@ -51,13 +51,21 @@ impl Default for ClientCodec {
     }
 }
 
+impl fmt::Debug for ClientCodec {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("h1::ClientCodec")
+            .field("flags", &self.inner.flags)
+            .finish_non_exhaustive()
+    }
+}
+
 impl ClientCodec {
     /// Create HTTP/1 codec.
     ///
     /// `keepalive_enabled` how response `connection` header get generated.
     pub fn new(config: ServiceConfig) -> Self {
         let flags = if config.keep_alive().enabled() {
-            Flags::KEEPALIVE_ENABLED
+            Flags::KEEP_ALIVE_ENABLED
         } else {
             Flags::empty()
         };
@@ -82,7 +90,7 @@ impl ClientCodec {
     }
 
     /// Check if last response is keep-alive
-    pub fn keepalive(&self) -> bool {
+    pub fn keep_alive(&self) -> bool {
         self.inner.conn_type == ConnectionType::KeepAlive
     }
 
@@ -105,7 +113,7 @@ impl ClientCodec {
 
 impl ClientPayloadCodec {
     /// Check if last response is keep-alive
-    pub fn keepalive(&self) -> bool {
+    pub fn keep_alive(&self) -> bool {
         self.inner.conn_type == ConnectionType::KeepAlive
     }
 
@@ -195,7 +203,7 @@ impl Encoder<Message<(RequestHeadType, BodySize)>> for ClientCodec {
                 // connection status
                 inner.conn_type = match head.as_ref().connection_type() {
                     ConnectionType::KeepAlive => {
-                        if inner.flags.contains(Flags::KEEPALIVE_ENABLED) {
+                        if inner.flags.contains(Flags::KEEP_ALIVE_ENABLED) {
                             ConnectionType::KeepAlive
                         } else {
                             ConnectionType::Close